new alert design added

+++ src/main/java/lk/ijse/gdse/controller/CustomerFormController.java
@@ -3,13 +3,10 @@ package lk.ijse.gdse.controller;
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXTextArea;
 import com.jfoenix.controls.JFXTextField;
-import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.SimpleBooleanProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import javafx.collections.transformation.SortedList;
-import javafx.concurrent.Task;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
@@ -21,14 +18,13 @@ import lk.ijse.gdse.dto.tm.CustomerTM;
 import lk.ijse.gdse.mail.Mail;
 import lk.ijse.gdse.model.CustomerModel;
 import lk.ijse.gdse.util.RegExPatterns;
+import lk.ijse.gdse.util.SystemAlert;
 import lk.ijse.gdse.util.TxtColours;
 
 import java.net.URL;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.*;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 
 public class CustomerFormController implements Initializable {
@@ -87,6 +83,9 @@ public class CustomerFormController implements Initializable {
     private ProgressBar progress;
 
     @FXML
+    private Label lblerror;
+
+    @FXML
     void addCustomerBtnOnAction(ActionEvent event) {
         boolean isExists = false;
         try {
@@ -99,10 +98,13 @@ public class CustomerFormController implements Initializable {
             if (!(txtCustomerId.getText().isEmpty() || txtCustomerName.getText().isEmpty() || txtGmail.getText().isEmpty() || txtContact.getText().isEmpty() || txtAddress.getText().isEmpty())) {
                 if (RegExPatterns.getCustomerId().matcher(txtCustomerId.getText()).matches()) {
                     TxtColours.setDefaultColours(txtCustomerId);
-                    if (RegExPatterns.getEmailPattern().matcher(txtGmail.getText()).matches()) {
-                        TxtColours.setDefaultColours(txtGmail);
-                        if (RegExPatterns.getContactPattern().matcher(txtContact.getText()).matches()) {
-                            TxtColours.setDefaultColours(txtContact);
+                    TxtColours.setDefaultColours(txtCustomerName);
+                    TxtColours.setDefaultColours(txtAddress);
+                    if (RegExPatterns.getContactPattern().matcher(txtContact.getText()).matches()) {
+                        TxtColours.setDefaultColours(txtContact);
+                        if (RegExPatterns.getEmailPattern().matcher(txtGmail.getText()).matches()) {
+                            TxtColours.setDefaultColours(txtGmail);
+                            lblerror.setText("");
 
                             String id = txtCustomerId.getText();
                             String name = txtCustomerName.getText();
@@ -113,34 +115,50 @@ public class CustomerFormController implements Initializable {
                             try {
                                 boolean isSaved = CustomerModel.addCustomer(new Customer(id, name, gmail, contact, address));
                                 if (isSaved) {
-                                    new Alert(Alert.AlertType.CONFIRMATION, "Customer saved :)").show();
+                                    new SystemAlert(Alert.AlertType.CONFIRMATION,"Confirmation","Customer saved!",ButtonType.OK).show();
                                     populateCustomerTable();
                                     searchFilter();
                                     clearTextfields();
                                 } else {
-                                    new Alert(Alert.AlertType.WARNING, "Customer not saved :(").show();
+                                    new SystemAlert(Alert.AlertType.WARNING,"Warning","Customer not saved!",ButtonType.OK).show();
                                 }
                             } catch (SQLException e) {
                                 e.printStackTrace();
-                                new Alert(Alert.AlertType.ERROR, "Something went wrong!").show();
+                                new SystemAlert(Alert.AlertType.ERROR,"Error","Somehing went wrong!",ButtonType.OK).show();
                             }
                         }else {
-                            new Alert(Alert.AlertType.WARNING,"Invalid contact number").show();
-                            TxtColours.setErrorColours(txtContact);
+                            lblerror.setText("Invalid gmail.");
+                            TxtColours.setErrorColours(txtGmail);
+
                         }
                     }else {
-                        new Alert(Alert.AlertType.WARNING,"Invalid gmail.").show();
-                        TxtColours.setErrorColours(txtGmail);
+                        lblerror.setText("Invalid contact number");
+                        TxtColours.setErrorColours(txtContact);
                     }
                 }else {
-                    new Alert(Alert.AlertType.WARNING,"Invalid customerId.").show();
+                    lblerror.setText("Invalid customerId.");
                     TxtColours.setErrorColours(txtCustomerId);
                 }
             }else {
-                new Alert(Alert.AlertType.WARNING,"Fill all the details").show();
+                lblerror.setText("Fill all the details");
+                if (txtCustomerId.getText().isEmpty()){
+                    TxtColours.setErrorColours(txtCustomerId);
+                }
+                if (txtCustomerName.getText().isEmpty()){
+                    TxtColours.setErrorColours(txtCustomerName);
+                }
+                if (txtAddress.getText().isEmpty()){
+                    TxtColours.setErrorColours(txtAddress);
+                }
+                if (txtContact.getText().isEmpty()){
+                    TxtColours.setErrorColours(txtContact);
+                }
+                if (txtGmail.getText().isEmpty()){
+                    TxtColours.setErrorColours(txtGmail);
+                }
             }
         }else {
-            new Alert(Alert.AlertType.WARNING,"Customer Id is already exists!").show();
+            lblerror.setText("Customer Id is already exists!");
             TxtColours.setErrorColours(txtCustomerId);
         }
     }
@@ -159,28 +177,35 @@ public class CustomerFormController implements Initializable {
         ButtonType yes = new ButtonType("Yes", ButtonBar.ButtonData.OK_DONE);
         ButtonType no = new ButtonType("No", ButtonBar.ButtonData.CANCEL_CLOSE);
 
-        Optional<ButtonType> result = new Alert(Alert.AlertType.INFORMATION,"Are you sure to delete ?",yes,no).showAndWait();
+        Optional<ButtonType> result = new SystemAlert(Alert.AlertType.INFORMATION,"Information","Are you sure to delete ?",yes,no).showAndWait();
 
-        if (result.orElse(no) == yes){
-                String id = txtCustomerId.getText();
-
+        if (result.orElse(no) == yes) {
+            String id = txtCustomerId.getText();
+            if (RegExPatterns.getCustomerId().matcher(id).matches()) {
+                TxtColours.setDefaultColours(txtCustomerId);
+                lblerror.setText("");
                 try {
                     boolean isDeleted = CustomerModel.deleteCustomer(id);
                     if (isDeleted) {
-                        new Alert(Alert.AlertType.CONFIRMATION, "Customer has deleted!").show();
+                        new SystemAlert(Alert.AlertType.CONFIRMATION,"Confirmation","Customer deleted!",ButtonType.OK).show();
                         clearTextfields();
                         populateCustomerTable();
                         searchFilter();
                     } else {
-                        new Alert(Alert.AlertType.WARNING, "Customer not deleted :(").show();
+                        new SystemAlert(Alert.AlertType.WARNING,"Warning","Customer not deleted!",ButtonType.OK).show();
                     }
                 } catch (SQLException e) {
                     e.printStackTrace();
-                    new Alert(Alert.AlertType.ERROR, "Something went wrong!").show();
+                    new SystemAlert(Alert.AlertType.ERROR,"Error","Somehing went wrong!",ButtonType.OK).show();
                 }
+            }else {
+                lblerror.setText("Invalid customer id");
+                TxtColours.setErrorColours(txtCustomerId);
             }
+        }
         }else {
-            new Alert(Alert.AlertType.WARNING,"Please give customerId").show();
+            lblerror.setText("Please enter the customer id");
+            TxtColours.setErrorColours(txtCustomerId);
         }
     }
 
@@ -190,10 +215,13 @@ public class CustomerFormController implements Initializable {
         if (!(txtCustomerId.getText().isEmpty() || txtCustomerName.getText().isEmpty() || txtGmail.getText().isEmpty() || txtContact.getText().isEmpty() || txtAddress.getText().isEmpty())) {
             if (RegExPatterns.getCustomerId().matcher(txtCustomerId.getText()).matches()) {
                 TxtColours.setDefaultColours(txtCustomerId);
-                if (RegExPatterns.getEmailPattern().matcher(txtGmail.getText()).matches()) {
-                    TxtColours.setDefaultColours(txtGmail);
-                    if (RegExPatterns.getContactPattern().matcher(txtContact.getText()).matches()) {
-                        TxtColours.setDefaultColours(txtContact);
+                TxtColours.setDefaultColours(txtCustomerName);
+                TxtColours.setDefaultColours(txtAddress);
+                if (RegExPatterns.getContactPattern().matcher(txtContact.getText()).matches()) {
+                    TxtColours.setDefaultColours(txtContact);
+                    if (RegExPatterns.getEmailPattern().matcher(txtGmail.getText()).matches()) {
+                        TxtColours.setDefaultColours(txtGmail);
+                        lblerror.setText("");
 
                         String id = txtCustomerId.getText();
                         String name = txtCustomerName.getText();
@@ -204,31 +232,47 @@ public class CustomerFormController implements Initializable {
                         try {
                             boolean isUpdated = CustomerModel.updateCustomer(new Customer(id, name, gmail, contact, address));
                             if (isUpdated) {
-                                new Alert(Alert.AlertType.CONFIRMATION, "Customer details updated!").show();
+                                new SystemAlert(Alert.AlertType.CONFIRMATION,"Confirmation","Customer updated!",ButtonType.OK).show();
                                 clearTextfields();
                                 populateCustomerTable();
                                 searchFilter();
                             } else {
-                                new Alert(Alert.AlertType.WARNING, "Customer details not updated!").show();
+                                new SystemAlert(Alert.AlertType.WARNING,"Warning","Customer not updated!",ButtonType.OK).show();
                             }
                         } catch (SQLException e) {
                             e.printStackTrace();
-                            new Alert(Alert.AlertType.ERROR, "Something went wrong").show();
+                            new SystemAlert(Alert.AlertType.ERROR,"Error","Somehing went wrong!",ButtonType.OK).show();
                         }
                     }else {
-                        new Alert(Alert.AlertType.WARNING,"Invalid contact number").show();
-                        TxtColours.setErrorColours(txtContact);
+                        lblerror.setText("Invalid gmail.");
+                        TxtColours.setErrorColours(txtGmail);
+
                     }
                 }else {
-                    new Alert(Alert.AlertType.WARNING,"Invalid gmail.").show();
-                    TxtColours.setErrorColours(txtGmail);
+                    lblerror.setText("Invalid contact number");
+                    TxtColours.setErrorColours(txtContact);
                 }
             }else {
-                new Alert(Alert.AlertType.WARNING,"Invalid customerId.").show();
+                lblerror.setText("Invalid customerId.");
                 TxtColours.setErrorColours(txtCustomerId);
             }
         }else {
-            new Alert(Alert.AlertType.WARNING,"Please fill all details").show();
+            lblerror.setText("Fill all the details");
+            if (txtCustomerId.getText().isEmpty()){
+                TxtColours.setErrorColours(txtCustomerId);
+            }
+            if (txtCustomerName.getText().isEmpty()){
+                TxtColours.setErrorColours(txtCustomerName);
+            }
+            if (txtAddress.getText().isEmpty()){
+                TxtColours.setErrorColours(txtAddress);
+            }
+            if (txtContact.getText().isEmpty()){
+                TxtColours.setErrorColours(txtContact);
+            }
+            if (txtGmail.getText().isEmpty()){
+                TxtColours.setErrorColours(txtGmail);
+            }
         }
     }
 
@@ -257,18 +301,18 @@ public class CustomerFormController implements Initializable {
             Thread thread = new Thread(mail);
             mail.valueProperty().addListener((a, oldValue, newValue) -> {
                 if (newValue){
-                    new Alert(Alert.AlertType.CONFIRMATION,"Announcement was sent.").show();
+                    new SystemAlert(Alert.AlertType.INFORMATION,"Email","Announcement was sent!",ButtonType.OK).show();
                     txtSubject.setText("");
                     txtMessage.setText("");
                 }else {
-                    new Alert(Alert.AlertType.WARNING,"Connection error").show();
+                    new SystemAlert(Alert.AlertType.NONE,"Connection Error","Connection Error!",ButtonType.OK).show();
                 }
             });
             progress.progressProperty().bind(mail.progressProperty());
             progress.visibleProperty().bind(mail.runningProperty());
             thread.start();
         }else {
-            new Alert(Alert.AlertType.WARNING,"Please fill the message!").show();
+            new SystemAlert(Alert.AlertType.WARNING,"Warning","Please fill the message!",ButtonType.OK).show();
         }
     }
 
@@ -321,7 +365,7 @@ public class CustomerFormController implements Initializable {
             ResultSet rs = CustomerModel.getAllCustomers();
                 data.clear();
                 while (rs.next()) {
-                    JFXButton button = new JFXButton("edit",new ImageView("F:\\1st semester final project\\moods salon\\src\\main\\resources\\img\\edit-97@30x.png"));
+                    JFXButton button = new JFXButton("edit",new ImageView("img/edit-97@30x.png"));
                     button.setContentDisplay(ContentDisplay.GRAPHIC_ONLY);
                     button.getStyleClass().add("infoBtn");
                     setButtonOnAction(button);
